# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Faydamed API Authentication",
        "description": "Authentication API endpoints for Faydamed system",
        "contact": {
            "email": "support@faydamed.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        },
        {
            "url": "https://faydaapi.dotprogrammers.com",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/api/cases/{id}": {
            "get": {
                "tags": [
                    "Cases"
                ],
                "summary": "Get case details",
                "description": "Retrieve full details of a specific case",
                "operationId": "getCaseById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Case not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Cases"
                ],
                "summary": "Update a case",
                "description": "Update details of an existing case",
                "operationId": "updateCase",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Case updated"
                    },
                    "404": {
                        "description": "Case not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cases"
                ],
                "summary": "Delete a case",
                "description": "Remove a case from the system",
                "operationId": "deleteCase",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Case deleted"
                    },
                    "404": {
                        "description": "Case not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/documents/{documentId}/ocr": {
            "post": {
                "tags": [
                    "OCR"
                ],
                "summary": "Run OCR on document",
                "description": "Process a PDF document through Google Vision to extract text content",
                "operationId": "processOcr",
                "parameters": [
                    {
                        "name": "documentId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OCR completed successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/documents/{documentId}/ocr/latest": {
            "get": {
                "tags": [
                    "OCR"
                ],
                "summary": "Get latest OCR result",
                "description": "Retrieve the most recent text extraction result for a specific document",
                "operationId": "getLatestOcr",
                "parameters": [
                    {
                        "name": "documentId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Document or result not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}